Version fonctionnelle

+++ src/utils/httpRequester.java
@@ -89,7 +89,7 @@ public final class httpRequester {
             while (i.hasNext()) {
                 JSONObject innerObj = (JSONObject) i.next();
                 int id_camion = Integer.parseInt((String) innerObj.get("id_camion"));
-                Coord position = new Coord(Integer.parseInt((String) innerObj.get("positionX_camion")), Integer.parseInt((String) innerObj.get("positionY_camion")));
+                Coord position = new Coord(Float.parseFloat((String) innerObj.get("positionX_camion")), Float.parseFloat((String) innerObj.get("positionY_camion")));
                 int id_caserne = Integer.parseInt((String) innerObj.get("id_caserne"));
                 int id_feu = Integer.parseInt((String) innerObj.get("id_feu"));
                 trucks.add(new Truck(id_camion, position, id_caserne, id_feu));
@@ -114,7 +114,7 @@ public final class httpRequester {
             while (i.hasNext()) {
                 JSONObject innerObj = (JSONObject) i.next();
                 int id_feu = Integer.parseInt((String) innerObj.get("id_feu"));
-                Coord position = new Coord(Integer.parseInt((String) innerObj.get("positionX")), Integer.parseInt((String) innerObj.get("positionY")));
+                Coord position = new Coord(Float.parseFloat((String) innerObj.get("positionX")), Float.parseFloat((String) innerObj.get("positionY")));
                 int intensite = Integer.parseInt((String) innerObj.get("intensite"));
                 if(intensite != 0)  {
                 	fires.add(new Fire(id_feu, position, intensite));
@@ -140,7 +140,7 @@ public final class httpRequester {
             while (i.hasNext()) {
                 JSONObject innerObj = (JSONObject) i.next();
                 int id_caserne = Integer.parseInt((String) innerObj.get("id_caserne"));
-                Coord position = new Coord(Integer.parseInt((String) innerObj.get("positionX_caserne")), Integer.parseInt((String) innerObj.get("positionY_caserne")));
+                Coord position = new Coord(Float.parseFloat((String) innerObj.get("positionX_caserne")), Float.parseFloat((String) innerObj.get("positionY_caserne")));
                 stations.add(new Station(id_caserne, position));
             }
 		} catch (Exception e) {
